+++ internal/circonus/submit.go
@@ -43,6 +43,7 @@ const (
 	traceTSFormat        = "20060102_150405.000000000"
 )
 
+// FlushCGM sends the tracking metrics collected in a CGM instance within the Check.
 func (c *Check) FlushCGM(ctx context.Context, ts *time.Time, lg zerolog.Logger, agentStats bool) {
 	if c.metrics != nil {
 		// TODO: add timestamp support to CGM (e.g. FlushMetricsWithTimestamp(ts))
@@ -74,15 +75,15 @@ func (c *Check) FlushCGM(ctx context.Context, ts *time.Time, lg zerolog.Logger, 
 		}
 
 		for {
-			submitCtx, submitCtxCancel := context.WithDeadline(ctx, time.Now().Add(30*time.Second))
-			err := c.SubmitMetrics(submitCtx, metrics, lg, !agentStats)
+			submitCtx, submitCtxCancel := context.WithDeadline(ctx, time.Now().Add(c.submitDeadline))
+			err := c.submitMetrics(submitCtx, metrics, lg, !agentStats)
 			if err == nil {
 				submitCtxCancel()
 				break
 			}
 
 			if errors.Is(err, context.DeadlineExceeded) {
-				c.log.Warn().Err(err).Msg("deadline reached submitting metrics, retrying")
+				c.log.Warn().Err(err).Str("deadline", c.submitDeadline.String()).Msg("deadline reached submitting metrics, retrying")
 				submitCtxCancel()
 				continue
 			}
@@ -95,8 +96,34 @@ func (c *Check) FlushCGM(ctx context.Context, ts *time.Time, lg zerolog.Logger, 
 	}
 }
 
-// Submit sends metrics to a circonus trap
-func (c *Check) SubmitMetrics(ctx context.Context, metrics map[string]MetricSample, resultLogger zerolog.Logger, includeStats bool) error {
+// FlushCollectorMetrics sends metrics from discrete collectors and sub-collectors
+func (c *Check) FlushCollectorMetrics(ctx context.Context, metrics map[string]MetricSample, resultLogger zerolog.Logger, includeStats bool) error {
+	var err error
+
+	for {
+		submitCtx, submitCtxCancel := context.WithDeadline(ctx, time.Now().Add(c.submitDeadline))
+		err = c.submitMetrics(submitCtx, metrics, resultLogger, includeStats)
+		if err == nil {
+			submitCtxCancel()
+			break
+		}
+
+		if errors.Is(err, context.DeadlineExceeded) {
+			c.log.Warn().Err(err).Str("deadline", c.submitDeadline.String()).Msg("deadline reached submitting metrics, retrying")
+			submitCtxCancel()
+			continue
+		}
+
+		submitCtxCancel()
+		c.log.Error().Err(err).Msg("submitting metrics")
+		break
+	}
+
+	return err
+}
+
+// submitMetrics does the heavy lifting to send metric batches to the trap check
+func (c *Check) submitMetrics(ctx context.Context, metrics map[string]MetricSample, resultLogger zerolog.Logger, includeStats bool) error {
 	if metrics == nil {
 		return errors.New("invalid metrics (nil)")
 	}
@@ -211,7 +238,7 @@ func (c *Check) SubmitMetrics(ctx context.Context, metrics map[string]MetricSamp
 		}
 	}
 
-	dataLen := subData.Len()
+	dataLen := len(rawData) // subData.Len()
 
 	reqStart := time.Now()
 
@@ -310,6 +337,7 @@ func (c *Check) SubmitMetrics(ctx context.Context, metrics map[string]MetricSamp
 		c.stats.RecvMetrics += result.Stats
 		c.stats.SentMetrics += uint64(len(metrics))
 		c.stats.SentBytes += uint64(dataLen)
+		c.stats.SentSize += uint64(subData.Len())
 		c.stats.BkrFiltered += result.Filtered
 		c.statsmu.Unlock()
 	}